Revert selectbox deletion
I accidentally deleted select box.css and box.js
in a previous commit. This re-adds the selectbox
files needed for front.html
and refactored the code so that the effects of
selectbox.css do not interfere with other files,
which was the reason I deleted it in the
first place.